Optimize code for better performance

+++ deep_draw/interface/utils.py
@@ -339,7 +339,7 @@ def image_from_dict(api_dict, dtype='uint8', encoding='utf-8'):
     return img
 
 
-def image_to_dict_RNN(image_array, dtype='float32', encoding='utf-8'):
+def image_to_dict_RNN(image_array, dtype='float16', encoding='utf-8'):
     '''
     Convert an ndarray representing a batch of images into a compressed string
     ----------
@@ -369,7 +369,7 @@ def image_to_dict_RNN(image_array, dtype='float32', encoding='utf-8'):
     return api_dict
 
 
-def image_from_dict_RNN(api_dict, dtype='float32', encoding='utf-8'):
+def image_from_dict_RNN(api_dict, dtype='float16', encoding='utf-8'):
     '''
     Convert an dict representing a batch of images into a ndarray
     ----------
